backend/requirements.txt
fastapi
uvicorn
fpdf
openpyxl
pillow
gspread
oauth2client
openai